feat: Add RunTaskWithMultiprocessing strategy
Same idea as https://github.com/getsentry/arroyo/pull/147.
The RunTaskWithMultiprocessing implementation is exactly the same
as ParallelTransform but it explicitly allows for side effects.
ParallelTransform is now simply an alias for RunTaskWithMultiprocessing.
