+++ arroyo/processing/strategies/__init__.py
@@ -7,7 +7,11 @@ from arroyo.processing.strategies.collect import CollectStep, ParallelCollectSte
 from arroyo.processing.strategies.commit import CommitOffsets
 from arroyo.processing.strategies.filter import FilterStep
 from arroyo.processing.strategies.produce import Produce
-from arroyo.processing.strategies.run_task import RunTask, RunTaskInThreads
+from arroyo.processing.strategies.run_task import (
+    RunTask,
+    RunTaskInThreads,
+    RunTaskWithMultiprocessing,
+)
 from arroyo.processing.strategies.transform import ParallelTransformStep, TransformStep
 
 __all__ = [
@@ -23,4 +27,5 @@ __all__ = [
     "Produce",
     "RunTask",
     "RunTaskInThreads",
+    "RunTaskWithMultiprocessing",
 ]
